Extract base class
We are having all sort of problems with this concern as we are getting
errors with private methods being invoked while running tests.

In any case, we are planning in breaking down this behaviour, so it
seems sensible to put it in a base class temporarily.
